requirements-8.txt
pyTelegramBotAPI==4.23.0
pymongo==4.10.1
Pillow==11.1.0
schedule==1.2.2
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
